+++ ircd/s_newconf.c
@@ -166,7 +166,7 @@ free_remote_conf(struct remote_conf *remote_p)
 
 void
 propagate_generic(struct Client *source_p, const char *command,
-		const char *target, int cap, const char *format, ...)
+		const char *target, uint64_t cap, const char *format, ...)
 {
 	char buffer[BUFSIZE];
 	va_list args;
@@ -185,7 +185,7 @@ propagate_generic(struct Client *source_p, const char *command,
 
 void
 cluster_generic(struct Client *source_p, const char *command,
-		int cltype, int cap, const char *format, ...)
+		int cltype, uint64_t cap, const char *format, ...)
 {
 	char buffer[BUFSIZE];
 	struct remote_conf *shared_p;
